move utils to utils.rs
